+++ modules/apps/document-library/document-library-repository-external-api/src/main/java/com/liferay/document/library/repository/external/model/ExtRepositoryFileEntryAdapter.java
@@ -290,7 +290,7 @@ public class ExtRepositoryFileEntryAdapter
 	}
 
 	@Override
-	public int getReadCount() {
+	public long getReadCount() {
 		return 0;
 	}
