+++ portal-kernel/src/com/liferay/portal/kernel/repository/model/FileEntryWrapper.java
@@ -241,7 +241,7 @@ public class FileEntryWrapper implements FileEntry, ModelWrapper<FileEntry> {
 	}
 
 	@Override
-	public int getReadCount() {
+	public long getReadCount() {
 		return _fileEntry.getReadCount();
 	}
